% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_transformations.R
\name{add_value}
\alias{add_value}
\title{Add a term to a variable}
\usage{
add_value(value, term)
}
\arguments{
\item{value}{The variable to which the term is added.}

\item{term}{The term to add.}
}
\value{
The variable with the term added.
}
\description{
Adds a specified term to a variable.
}
