% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_transformations.R
\name{center_variable}
\alias{center_variable}
\title{Center a variable}
\usage{
center_variable(x, center)
}
\arguments{
\item{x}{The variable to center.}

\item{center}{The centering value.}
}
\value{
The centered variable.
}
\description{
Subtracts a centering value from a variable.
}
\examples{
center_variable(10, 5) # Returns 5
center_variable(c(1, 2, 3, 4, 5), 3) # Returns c(-2, -1, 0, 1, 2)
}
