% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffview.R
\name{.format_diff_output}
\alias{.format_diff_output}
\title{Internal function to format diff output}
\usage{
.format_diff_output(differences, config_path1, config_path2)
}
\arguments{
\item{differences}{List of differences from comparison}

\item{config_path1}{Path to first config file}

\item{config_path2}{Path to second config file}
}
\value{
Character vector with formatted diff output
}
\description{
Internal function to format diff output
}
