% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffview.R
\name{.format_value}
\alias{.format_value}
\title{Internal function to format values for display}
\usage{
.format_value(value)
}
\arguments{
\item{value}{Value to format}
}
\value{
Character representation of the value
}
\description{
Internal function to format values for display
}
