% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_transformations.R
\name{exp_transform}
\alias{exp_transform}
\title{Exponential-transform a variable}
\usage{
exp_transform(x, base = exp(1))
}
\arguments{
\item{x}{The variable to exponentiate.}

\item{base}{The base of the exponentiation. Defaults to \code{exp(1)} (natural exponentiation).}
}
\value{
The exponentiated variable.
}
\description{
Exponentiates a variable.
}
\examples{
exp_transform(2, base = 10) # Returns 100
exp_transform(2) # Returns exp(2)
}
