% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_config}
\alias{load_config}
\title{Load a Rydra configuration file}
\usage{
load_config(file_path)
}
\arguments{
\item{file_path}{Path to the YAML configuration file.}
}
\value{
A list representing the parsed YAML configuration.
}
\description{
This function reads a YAML file and returns it as an R list.
}
\examples{
\dontrun{
  config <- load_config("path/to/your/config.yaml")
}
}
