% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_transformations.R
\name{log_transform}
\alias{log_transform}
\title{Log-transform a variable}
\usage{
log_transform(x, base = exp(1))
}
\arguments{
\item{x}{The variable to log-transform.}

\item{base}{The base of the logarithm.}
}
\value{
The log-transformed variable.
}
\description{
Log-transforms a variable.
}
