% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_transformations.R
\name{multiply_by}
\alias{multiply_by}
\title{Multiply a variable by a multiplier}
\usage{
multiply_by(value, multiplier)
}
\arguments{
\item{value}{The variable to multiply.}

\item{multiplier}{The multiplier value.}
}
\value{
The variable multiplied by the multiplier.
}
\description{
Multiplies a variable by a specified multiplier.
}
\examples{
multiply_by(10, 5) # Returns 50
}
