% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_transformations.R
\name{square_variable}
\alias{square_variable}
\title{Square a variable}
\usage{
square_variable(x)
}
\arguments{
\item{x}{The variable to square.}
}
\value{
The squared variable.
}
\description{
Squares a variable.
}
