% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_transformations.R
\name{truncate_variable}
\alias{truncate_variable}
\title{Truncate a variable}
\usage{
truncate_variable(x, min_val, max_val)
}
\arguments{
\item{x}{The variable to truncate.}

\item{min_val}{The minimum value.}

\item{max_val}{The maximum value.}
}
\value{
The truncated variable.
}
\description{
Truncates a variable to a specified minimum and maximum range.
Values less than min_val are set to min_val.
Values greater than max_val are set to max_val.
}
