% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{apply_transformations}
\alias{apply_transformations}
\title{Apply transformations to data based on model and global configurations}
\usage{
apply_transformations(
  model_yaml_config,
  data,
  transformation_R_functions,
  full_config
)
}
\arguments{
\item{model_yaml_config}{The specific configuration block for the current model
(e.g., \code{config[[model_name]]}), which should contain a \code{transformations} list.}

\item{data}{A data frame or list containing the input data.}

\item{transformation_R_functions}{A named list of R functions available for transformations
(e.g., base functions, user-defined functions).}

\item{full_config}{The full top-level configuration list, used to access
global elements like \code{centering}.}
}
\value{
A data frame with the transformed variables added.
}
\description{
This function applies a series of transformations to the input data as
defined in a specific model's configuration block within a larger YAML structure.
}
\keyword{internal}
