% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_transformations.R
\name{center_variable}
\alias{center_variable}
\title{Center a variable}
\usage{
center_variable(x, center)
}
\arguments{
\item{x}{The variable to center.}

\item{center}{The centering value.}
}
\value{
The centered variable.
}
\description{
Subtracts a centering value from a variable.
}
