% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_transformations.R
\name{exp_transform}
\alias{exp_transform}
\title{Exponential-transform a variable}
\usage{
exp_transform(x, base = exp(1))
}
\arguments{
\item{x}{The variable to exponentiate.}

\item{base}{The base of the exponentiation (default: exp(1)).}
}
\value{
The exponentiated variable.
}
\description{
Exponentiates a variable.
}
